+++ javascript/node/selenium-webdriver/lib/http.js
@@ -429,34 +429,29 @@ class Executor {
       return doSend(this, request).then(response => {
         this.log_.finer(() => `>>>\n${request}\n<<<\n${response}`);
 
-        let parsed =
-            parseHttpResponse(
-                command, /** @type {!Response} */ (response), this.w3c);
+        let httpResponse = /** @type {!Response} */(response);
+        let {isW3C, value} = parseHttpResponse(command, httpResponse);
 
         if (command.getName() === cmd.Name.NEW_SESSION
             || command.getName() === cmd.Name.DESCRIBE_SESSION) {
-          if (!parsed || !parsed['sessionId']) {
+          if (!value || !value.sessionId) {
             throw new error.WebDriverError(
-                'Unable to parse new session response: ' + response.body);
+                `Unable to parse new session response: ${response.body}`);
           }
 
           // The remote end is a W3C compliant server if there is no `status`
           // field in the response. This is not applicable for the DESCRIBE_SESSION
           // command, which is not defined in the W3C spec.
           if (command.getName() === cmd.Name.NEW_SESSION) {
-            this.w3c = this.w3c || !('status' in parsed);
+            this.w3c = this.w3c || isW3C;
           }
 
-          return new Session(parsed['sessionId'], parsed['value']);
+          // No implementations use the `capabilities` key yet...
+          let capabilities = value.capabilities || value.value;
+          return new Session(value.sessionId, capabilities);
         }
 
-        if (parsed
-            && typeof parsed === 'object'
-            && 'value' in parsed) {
-          let value = parsed['value'];
-          return typeof value === 'undefined' ? null : value;
-        }
-        return parsed;
+        return typeof value === 'undefined' ? null : value;
       });
     });
   }
@@ -482,40 +477,45 @@ function tryParse(str) {
  *
  * @param {!cmd.Command} command The command the response is for.
  * @param {!Response} httpResponse The HTTP response to parse.
- * @param {boolean} w3c Whether the response should be processed using the
- *     W3C wire protocol.
- * @return {?} The parsed response.
+ * @return {{isW3C: boolean, value: ?}} An object describing the parsed
+ *     response. This object will have two fields: `isW3C` indicates whether
+ *     the response looks like it came from a remote end that conforms with the
+ *     W3C WebDriver spec, and `value`, the actual response value.
  * @throws {WebDriverError} If the HTTP response is an error.
  */
-function parseHttpResponse(command, httpResponse, w3c) {
+function parseHttpResponse(command, httpResponse) {
+  if (httpResponse.status < 200) {
+    // This should never happen, but throw the raw response so users report it.
+    throw new error.WebDriverError(
+        `Unexpected HTTP response:\n${httpResponse}`);
+  }
+
   let parsed = tryParse(httpResponse.body);
-  if (parsed !== undefined) {
-    if (httpResponse.status < 200) {
-      // This should never happen, but throw the raw response so
-      // users report it.
-      throw new error.WebDriverError(
-          `Unexpected HTTP response:\n${httpResponse}`);
-    }
+  if (parsed && typeof parsed === 'object') {
+    let value = parsed.value;
+    let isW3C =
+        value !== null && typeof value === 'object'
+            && typeof parsed.status === 'undefined';
 
-    if (w3c) {
-      if (httpResponse.status > 399) {
-        error.throwDecodedError(parsed);
+    if (!isW3C) {
+      error.checkLegacyResponse(parsed);
+
+      // Adjust legacy new session responses to look like W3C to simplify
+      // later processing.
+      if (command.getName() === cmd.Name.NEW_SESSION
+          || command.getName() == cmd.Name.DESCRIBE_SESSION) {
+        value = parsed;
       }
-      return parsed;
+
+    } else if (httpResponse.status > 399) {
+      error.throwDecodedError(value);
     }
 
-    // If this is a new session command, we need to check for a W3C compliant
-    // error object. This is necessary since a successful new session command
-    // is what puts the executor into W3C mode.
-    if (httpResponse.status > 399
-        && (command.getName() == cmd.Name.NEW_SESSION
-            || command.getName() === cmd.Name.DESCRIBE_SESSION)
-        && error.isErrorResponse(parsed)) {
-      error.throwDecodedError(parsed);
-    }
+    return {isW3C, value};
+  }
 
-    error.checkLegacyResponse(parsed);
-    return parsed;
+  if (parsed !== undefined) {
+    return {isW3C: false, value: parsed};
   }
 
   let value = httpResponse.body.replace(/\r\n/g, '\n');
@@ -528,7 +528,7 @@ function parseHttpResponse(command, httpResponse, w3c) {
     throw new error.WebDriverError(value);
   }
 
-  return value || null;
+  return {isW3C: false, value: value || null};
 }